Updated select-types..., use-windows-control...ig, and configure-authorized-apps...
select-types:
1) removed the ConfigCI section, including the SKUs blurb
2) Removed the Supplemental policies table and blurb.
3) Added Supplemental policies info to table 1, as well as info to the table 1 text.

configure-authorized-apps... and use-windows-defender-application...
Added fsutil information as penultimate section in both files.

Note: I made these changes yesterday to a different branch, committed and pushed to origin, but I wasn't able to create a PR. It said that there were not changes in the branch to compare to public. So, I deleted that branch and started anew.