Replaced "**" with "<b>" to resolve rendering the asterisks as body text
I don't know why, in this one place, the build rendered the asterisks around the name of a policy instead of making the name bold. There is often a problem with using Markdown and HTML together in the same line, but this particular patterns works elsewhere in this file.

This commit replaces the asterisks with HTML bold tags.