Moved HR to see if that will allow correct processing of "**ADMX_pca/DetectDeprecatedComponentFailuresPolicy**"
The HR tag that precedes "**ADMX_pca/DetectDeprecatedComponentFailuresPolicy**" is closer than other similar instances in this article. In preview, "**ADMX_pca/DetectDeprecatedComponentFailuresPolicy**" was rendered as plain text with asterisks around it, rather than bold text as is the case in the other similar instances. Markdown is unreliable when in or near HTML, so another solution here would have been to replace the asterisks HTML bold tags.