Fix "find/replace" accident
windows-1...overview.md
https://microsoft-ce-csi.acrolinx.cloud/api/v1/checking/scorecards/67bf32c7-d246-49a1-9603-531cf8904857#CORRECTNESS

Line 140: Group Policy requires that the you have AD DS and that the Windows 10/11 Enterprise devices  (delete "the" between "that" and "you")

Line 140: are joined to the your AD DS domain. (delete "the" between "to" and "your")

Line 146: The primary App-V components that the you must have are as follows: (delete "the" between "that" and "you")

Line 161: client-side components that you you’ll need to download, activate, and install. (delete "you" between "that" and "you'll)