Update prepare-your-organization-for-windows-to-go.md
Line 3: the the > the
Line 95: If you configure Windows To Go drives for scenarios where drives may remain unused for extended period of time such as use in continuance of operations scenarios, the AD DS computer account objects that correspond to Windows To Go drives have the potential to become stale and be pruned during maintenance operations. > If you configure Windows To Go drives for scenarios where drives may remain unused for extended periods of time such as used in continuance of operations scenarios, the AD DS computer account objects that correspond to Windows To Go drives have the potential to become stale and be pruned during maintenance operations. 
Line 95: clean up > clean
Lines 29, 36,59, 78, and 99: is not > isn't
Lines 75 and 89: they are > they're
Line 78: those computers which have the ability to > those computers that can 
Lines 50, 61 (x2), and 82,  : additional > other
Line 80: you are > you're
Line 82: ; > :