Update mcc-isp-create-provision-deploy.md
The configuration of settings like storage and routing is actually different than the provisioning of the server itself. Modified a few sentences after reviewing the live code and experience and comparing the text in the documentation.

Removed IoT Central as there isn't anything from IoT Central that is directly installed on the server and it should be completely transparent to the customer.

Added information about Docker as this is installed.