Clarification of Settings Catalog options
We currently list both a method of configuring this with Settings Catalog, and with a Custom profile. Arguably the custom profile should just go these days, but in any case the Settings Catalog entry was incomplete and needed clarification on the Telemetry options, as we did not state recommended settings. Those settings are also not required, but recommended, so I have moved them as such.

I've also added brief wording to clarify that you need not do a Settings Catalog AND a custom profile, as this has been misunderstood before.