Clarify use of Device Credentials for Azure Virtual Desktop
The NOTE section regarding User vs Device credentials and its reference to Azure Virtual Desktop support is misleading.  Updated the note with a link to specify that Device Credential is only supported for MULTI-SESSION host pools, with a link for reference, and User Credential is still supported for personal host pools.