Add documentation for ScanBeforeInitialLogon
Adding documentation for a registry value introduced being introduced in Win11 22H2 to support scanning for and installing updates before a first user logon.  This is meant to be used only for devices (such as CloudPCs based on Azure VM) where there is a significant delay between the completion of the first user experience and the initial logon.  Included a warning that using this value may have a negative effect when used in retail devices.