In getting-started-with-the-user-state-migration-tool.md, say to use the "/genconfig option" instead of the "ScanState Syntax option"
The part that said, "use the ScanState Syntax option" didn't make sense, so--as you can probably see in the diff--I replaced that part with "use the /genconfig option" in order to make that part make sense.

I also changed the link that's in that part to link directly to the "Migration rule options" section of the "ScanState syntax" article instead of the top of that article because that seems more helpful than linking to the top of the article. However, if it's Microsoft's style to link to the top of an article instead of directly to a relevant section, feel free to revert that change in order to be consistent with your prescribed style.