Update lock-down-windows-11-to-specific-apps.md
Commented out the coming soon section as it is causing customer issues
Updated the WMI Bridge step by step instructions to remove ambiguity
Updated the Sample PowerShell and XML snip-its since the existing samples were broken.

I would really like to refactor this topic into a quickstart that leverages a canned sample with a separate XML reference that is combined with the existing XML reference topic, but that is going to require more time.