Update windows-autopatch-faq.yml
Update broken anchors. The anchor for pausing quality updates is #pause-and-resume-a-release (https://learn.microsoft.com/en-us/windows/deployment/windows-autopatch/operate/windows-autopatch-groups-windows-quality-update-overview#pause-and-resume-a-release), not pausing-and-resuming-a-release. The content for pause/resume feature updates has moved to a different page (https://learn.microsoft.com/en-us/windows/deployment/windows-autopatch/operate/windows-autopatch-groups-manage-windows-feature-update-release#pause-and-resume-a-release) with a different anchor.