Update deploy-wdac-policies-with-script.md
Removed a stipulation which implies that signed WDAC policies have to be placed within System32 and EFI locations. 
In many cases they should ONLY be placed in the EFI partition. (NOT the System32 location.)
This updated wording matches the behavior of the CiTool (when using `CiTool --update-policies` to deploy a new signed policy).

NOTE: It's recommended that the wording be refined even further to emphasize that you **SHOULDN'T** place a signed policy in both locations. We are aware of blue-screens affecting Windows 11 devices which have signed policies in both locations. (Affected models: Dell Precision 3680, Dell Precision 3650 Tower, Dell OptiPlex Micro 7010, Dell Inspiron 15 3511.) It's possible others could be affected.